[TASK] Use NPM dependencies from outside Docker.
This means that in order to build the Docker image, one must first run a clean npm install. There can be no symlinked or missing packages otherwise the resulting Docker image is corrupt.

With our current workflow where Circle is the only thing that ever builds the Docker image - this should be fine.