BREAKING: Switch to PBKDF2
Removes password hashing from five-bells-ledger and uses the facility in five-bells-shared.

Makes password hashing asynchronous. Each password hashing iteration with 10 rounds of bcrypt takes around 60ms in my tests. This is for every request to the ledger and on the main thread. With this patch, the password hashing becomes asynchronous and is removed from the main thread. This reduces the total duration of the unit tests from 35s to about 8s on my machine. The number of "prepare transfer" requests per second increases from 60 req/s to about 110 req/s.

Note that the unit tests were using some rudimentary partial caching for password hashes, so the impact on a live system could be even larger, depending on the number of cores available.

This change does not modify the public API, however it does change the database schema and password hashing and therefore necessitates a database reset.