fix: Check expiry date before fulfilling a transfer
Before fulfilling a transfer, the ledger did not check if the transfer was expired. This has several

securtiy-critical consequences, e.g.: An attacker can bring a connector to put funds on hold and

trick the ledger to fulfill the hold after it is expired. In this scenario, the connector can not

pass on the fulfillment in time to the source ledger and, thus, losses the funds.