fix: only send notifications once per listener
Currently, if a single websocket connection is listening to multiple accounts and more than one of those accounts is involved in a transfer, the transfer notification will be transmitted multiple times. This patch changes the behavior such that notifications are only emitted once per listener which means once per websocket connection.

Fixes #385