build: move sqlite3 from optionalDependencies to devDependencies
As described in https://docs.npmjs.com/files/package.json#optionaldependencies, a package should
only be in optionalDependencies if the code can handle both its presence and its absence. This is
not the case for sqlite3 when running the tests, so optionalDependencies is the wrong place for it.
We also don't need sqlite3 in production, and if you use five-bells-ledger as a dependency, then the
way you use it will dictate which database engine you want to use, and your package would depend on
that separately, so having it as a full dependency is also not necessary. Therefore, it needs to be
at least in devDependencies.